requirements.txt
torch
torchvision
transformers
datasets
opencv-python
Pillow
tqdm
scikit-learn
gradio
clip @ git+https://github.com/openai/CLIP.git
